small and big zk program for cache regression
