+++ tests/perf-regression/perf-regression.sh
@@ -20,7 +20,7 @@ MODE=$1
 ./run src/examples/zkprogram/gadgets.ts --bundle "$MODE"
 ./run src/examples/zkprogram/side-loading/run.ts --bundle "$MODE"
 ./run src/examples/zkprogram/runtime-table/run.ts --bundle "$MODE"
-./run src/examples/zkprogram/program-small-big.ts --bundle "$MODE"
+./run src/examples/zkprogram/small-big/run.ts --bundle "$MODE"
 
 # Run CS + zkApps performance regression tests
 ./run tests/perf-regression/perf-regression.ts --bundle "$MODE"
